# gin-petclinic-service/docs/swagger.yaml
basePath: /v1
definitions:
  errors.ErrorResponse:
    properties:
      details: {}
      message:
        example: status bad request
        type: string
      status:
        example: 400
        type: integer
    type: object
  model.Context:
    properties:
      count:
        type: integer
    type: object
  owner.AddRequest:
    properties:
      address:
        type: string
      city:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      telephone:
        type: string
    type: object
  owner.Response:
    properties:
      address:
        type: string
      city:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      pets:
        items:
          $ref: '#/definitions/pet.Response'
        type: array
      telephone:
        type: string
    type: object
  owner.Responses:
    properties:
      context:
        $ref: '#/definitions/model.Context'
      owners:
        items:
          $ref: '#/definitions/owner.Response'
        type: array
    type: object
  owner.UpdateRequest:
    properties:
      address:
        type: string
      city:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      telephone:
        type: string
    type: object
  owner.UpdateResponse:
    properties:
      address:
        type: string
      city:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      telephone:
        type: string
    type: object
  pet.AddRequest:
    properties:
      birthdate:
        type: string
      name:
        type: string
      ownerID:
        type: integer
      typeID:
        type: integer
    required:
    - birthdate
    - name
    - ownerID
    - typeID
    type: object
  pet.Response:
    properties:
      birthdate:
        type: string
      id:
        type: integer
      name:
        type: string
      type:
        type: string
      visits:
        items:
          $ref: '#/definitions/visit.Response'
        type: array
    type: object
  pet.Responses:
    properties:
      context:
        $ref: '#/definitions/model.Context'
      pets:
        items:
          $ref: '#/definitions/pet.Response'
        type: array
    type: object
  pet.UpdateRequest:
    properties:
      birthdate:
        type: string
      id:
        type: integer
      name:
        type: string
      ownerID:
        type: integer
      typeID:
        type: integer
    required:
    - birthdate
    - id
    - name
    - ownerID
    - typeID
    type: object
...
  visit.Response:
    properties:
      description:
        type: string
      id:
        type: integer
      visitDate:
        type: string
    type: object
...
host: localhost:8443
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a pet clinic API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pet Clinic API
  version: "1.0"
paths:
  /owners:
    get:
      description: Get owners by last name
      parameters:
      - description: Owner last name
        in: query
        name: last-name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/owner.Responses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Search owners by last name
      tags:
      - owners
    post:
      description: Insert new owner
      parameters:
      - description: Add owner
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/owner.AddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/owner.UpdateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Insert new owner
      tags:
      - owners
  /owners/{id}:
    get:
      description: Get owner by ID
      parameters:
      - description: Owner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/owner.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Search owner by ID
      tags:
      - owners
    put:
      description: Update new owner
      parameters:
      - description: Owner ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update owner
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/owner.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/owner.UpdateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update owner
      tags:
      - owners
  /owners/{id}/pets:
    get:
      description: Get owner that has pets by ID
      parameters:
      - description: Owner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/owner.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Search owner that has pets by ID
      tags:
      - owners
  /owners/all:
    get:
      description: Get all owners
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/owner.Responses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List all owners
      tags:
      - owners
  /pets:
    get:
      description: Get pet by name
      parameters:
      - description: Pet Name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pet.Responses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get pet by name
      tags:
      - pets
    post:
      description: Add a new pet
      parameters:
      - description: Add pet
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/pet.AddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pet.Responses'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Add a new pet
      tags:
      - pets
  /pets/{id}:
    get:
      description: Get pet by ID
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pet.Responses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get pet by ID
      tags:
      - pets
    put:
      description: update pet
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update pet
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/pet.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pet.Responses'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: update a pet
      tags:
      - pets
  /pets/{id}/visits:
    get:
      description: Get pet with visits by ID
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pet.Responses'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get pet with visits by ID
      tags:
      - pets
  /pets/all:
    get:
      description: Get all pets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pet.Responses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List all pets
      tags:
      - pets
...
schemes:
- https
swagger: "2.0"