# gin-petclinic-service/internal/api/owner/owner_request.go
package owner

import (
	"github.com/rhtran/gin-petclinic-service/pkg/infra/repository"
	"github.com/rhtran/gin-petclinic-service/pkg/model"
)

type AddRequest struct {
	FirstName string `json:"firstName"`
	LastName  string `json:"lastName"`
	Address   string `json:"address"`
	City      string `json:"city"`
	Telephone string `json:"telephone"`
}

type UpdateRequest struct {
	Id        uint   `json:"id"`
	FirstName string `json:"firstName"`
	LastName  string `json:"lastName"`
	Address   string `json:"address"`
	City      string `json:"city"`
	Telephone string `json:"telephone"`
}

func ToOwnerEntity(ownerRequest *AddRequest) *repository.Owner {
	return &repository.Owner{
		Person: model.Person{
			FirstName: ownerRequest.FirstName,
			LastName:  ownerRequest.LastName,
		},

		Address:   ownerRequest.Address,
		City:      ownerRequest.City,
		Telephone: ownerRequest.Telephone,
	}
}

func ToOwnerEntityFromUpdateRequest(ownerRequest *UpdateRequest) *repository.Owner {
	return &repository.Owner{
		Person: model.Person{
			FirstName: ownerRequest.FirstName,
			LastName:  ownerRequest.LastName,
		},

		Address:   ownerRequest.Address,
		City:      ownerRequest.City,
		Telephone: ownerRequest.Telephone,
	}
}